+++ gusto_embedded/RELEASES.md
@@ -369,3 +369,13 @@ Based on:
 - [typescript v0.7.0] gusto_embedded
 ### Releases
 - [NPM v0.7.0] https://www.npmjs.com/package/@gusto/embedded-api/v/0.7.0 - gusto_embedded
+
+## 2025-10-01 00:15:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.631.5 (2.716.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.1] gusto_embedded
+### Releases
+- [NPM v0.7.1] https://www.npmjs.com/package/@gusto/embedded-api/v/0.7.1 - gusto_embedded